Rewrote basic class structure
Spend a while iteratively designing the basic foundational structure of the code and how it flows.

My theory being if i spend time on strong foundations it will speed up and simplify development later and will help efficiently find and fix bugs.

Logical code flow and readable code are the two goals with this.